+++ bw.workflow.commons/src/main/java/sa/gov/nic/bio/bw/workflow/commons/beans/PersonInfo.java
@@ -7,21 +7,21 @@ import java.util.Date;
 
 public class PersonInfo extends JavaBean
 {
-	private long samisId;
+	private Long samisId;
 	private Name name;
 	private Date birthDate;
 	private String birthPlace;
 	private String birthHijGregInd;
 	private String personType;
-	private int nationality;
+	private Integer nationality;
 	private String nationalityDesc;
-	private int gender;
+	private Integer gender;
 	private PersonIdInfo identityInfo;
 	private String face;
 	private Boolean isOut;
 	
-	public long getSamisId(){return samisId;}
-	public void setSamisId(long samisId){this.samisId = samisId;}
+	public Long getSamisId(){return samisId;}
+	public void setSamisId(Long samisId){this.samisId = samisId;}
 	
 	public Name getName(){return name;}
 	public void setName(Name name){this.name = name;}
@@ -38,14 +38,14 @@ public class PersonInfo extends JavaBean
 	public String getPersonType(){return personType;}
 	public void setPersonType(String personType){this.personType = personType;}
 	
-	public int getNationality(){return nationality;}
-	public void setNationality(int nationality){this.nationality = nationality;}
+	public Integer getNationality(){return nationality;}
+	public void setNationality(Integer nationality){this.nationality = nationality;}
 	
 	public String getNationalityDesc(){return nationalityDesc;}
 	public void setNationalityDesc(String nationalityDesc){this.nationalityDesc = nationalityDesc;}
 	
-	public int getGender(){return gender;}
-	public void setGender(int gender){this.gender = gender;}
+	public Integer getGender(){return gender;}
+	public void setGender(Integer gender){this.gender = gender;}
 	
 	public PersonIdInfo getIdentityInfo(){return identityInfo;}
 	public void setIdentityInfo(PersonIdInfo identityInfo){this.identityInfo = identityInfo;}