+++ bw.workflow.registerconvictednotpresent/src/main/java/module-info.java
@@ -9,4 +9,9 @@ module bw.workflow.registerconvictednotpresent
 	opens sa.gov.nic.bio.bw.workflow.registerconvictednotpresent to bw.core;
 	opens sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.fxml to bw.core;
 	opens sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.controllers to javafx.fxml, bw.core;
+	opens sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.tasks to bw.core;
+	opens sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.beans to org.apache.commons.lang3;
+	exports sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.webservice;
+	exports sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.tasks;
+	exports sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.beans;
 }