+++ bw.workflow.registerconvictednotpresent/src/main/java/sa/gov/nic/bio/bw/workflow/registerconvictednotpresent/utils/RegisterConvictedNotPresentErrorCodes.java
@@ -2,7 +2,7 @@ package sa.gov.nic.bio.bw.workflow.registerconvictednotpresent.utils;
 
 public enum RegisterConvictedNotPresentErrorCodes
 {
-	C009_00001;
+	C009_00001, C009_00002, C009_00003;
 	
 	public final String getCode()
 	{